+++ docs/src/examples/components/Label/Variations/LabelExampleIconPosition.shorthand.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { Label } from '@stardust-ui/react'
+
+const LabelExampleIconPositionShorthand = () => (
+  <Label circular icon="close" iconPosition="end" content="Label with icon after the content" />
+)
+
+export default LabelExampleIconPositionShorthand