+++ docs/src/examples/components/Label/Variations/index.tsx
@@ -9,6 +9,26 @@ const Variations = () => (
       description="A label can be circular"
       examplePath="components/Label/Variations/LabelExampleCircular"
     />
+    <ComponentExample
+      title="Icon"
+      description="The label can contain an icon"
+      examplePath="components/Label/Variations/LabelExampleIcon"
+    />
+    <ComponentExample
+      title="Icon position"
+      description="The icon inside the label can be aligned before or after the content"
+      examplePath="components/Label/Variations/LabelExampleIconPosition"
+    />
+    <ComponentExample
+      title="Clickable icon"
+      description="The icon inside the label can be clickable"
+      examplePath="components/Label/Variations/LabelExampleOnIconClick"
+    />
+    <ComponentExample
+      title="The Icon inside Label can be customized"
+      description="The Icon component inside the Label can be defined with customizing it's prop"
+      examplePath="components/Label/Variations/LabelExampleIconAsShorthand"
+    />
   </ExampleSection>
 )