+++ src/components/Icon/Icon.tsx
@@ -1,6 +1,6 @@
 import React, { CSSProperties } from 'react'
 import PropTypes from 'prop-types'
-import { customPropTypes, UIComponent, SUI } from '../../lib'
+import { customPropTypes, UIComponent, SUI, createShorthandFactory } from '../../lib'
 
 import iconRules from './iconRules'
 import iconVariables from './iconVariables'
@@ -38,9 +38,11 @@ export interface IconProps {
   xSpacing?: IconXSpacing
   style?: CSSProperties
   title?: string
+  variables?: (siteVariables: object) => object
 }
 
 class Icon extends UIComponent<IconProps, {}> {
+  static create: Function
   static className = 'ui-icon'
 
   static displayName = 'Icon'
@@ -118,5 +120,7 @@ class Icon extends UIComponent<IconProps, {}> {
     return <ElementType className={classes.root} {...rest} />
   }
 }
+
+Icon.create = createShorthandFactory(Icon, name => ({ name }))
 
 export default Icon